tests/RESTful/requirements.txt
requests
nose
nose-testconfig
nose-parameterized>=0.5.0